🚚 Re-organize some classes for clarity
