+++ CodeMirror6/CodeMirror6Wrapper.razor.cs
@@ -1,3 +1,4 @@
+using System.Collections.ObjectModel;
 using CodeMirror6.Models;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
@@ -97,7 +98,7 @@ public partial class CodeMirror6Wrapper : ComponentBase, IAsyncDisposable
     /// The active markdown styles at the current selection(s)
     /// </summary>
     /// <value></value>
-    [Parameter] public EventCallback<List<string>> MarkdownStylesAtSelectionsChanged { get; set; }
+    [Parameter] public EventCallback<ReadOnlyCollection<string>> MarkdownStylesAtSelectionsChanged { get; set; }
     /// <summary>
     /// Whether to allow vertical resizing similar to a textarea
     /// </summary>
@@ -150,16 +151,6 @@ public partial class CodeMirror6Wrapper : ComponentBase, IAsyncDisposable
     public CMCommandDispatcher? Commands => CmJsInterop?.Commands;
 
     private string LoadingDivId => $"{Id}_Loading";
-
-    /// <summary>
-    /// JavaScript interop instance
-    /// </summary>
-    internal CodeMirrorJsInterop? CmJsInterop = null;
-    internal CodeMirrorConfiguration Config = null!;
-    internal CodeMirrorState State = new();
-
-    private bool hasFocus;
-    private bool shouldRender = true;
     private string ResizeStyle => AllowVerticalResize && AllowHorizontalResize
         ? "both"
         : AllowVerticalResize ? "vertical"
@@ -167,105 +158,11 @@ public partial class CodeMirror6Wrapper : ComponentBase, IAsyncDisposable
         : "none";
 
     /// <summary>
-    /// The document contents has changed
+    /// JavaScript interop instance
     /// </summary>
-    /// <param name="value"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task DocChangedFromJS(string value)
-    {
-        if (Doc?.Replace("\r", "") == value?.Replace("\r", "")) return;
-        Doc = value?.Replace("\r", "") ?? "";
-        Config.Doc = Doc;
-        await DocChanged.InvokeAsync(Doc);
-    }
-
-    /// <summary>
-    /// The document focus has changed
-    /// </summary>
-    /// <param name="value"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task FocusChangedFromJS(bool value)
-    {
-        if (hasFocus == value) return;
-        hasFocus = value;
-        await FocusChanged.InvokeAsync(hasFocus);
-        if (!hasFocus) await SelectionChanged.InvokeAsync(null);
-    }
-
-    /// <summary>
-    /// The cursor position or selections have changed
-    /// </summary>
-    /// <param name="values"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task SelectionSetFromJS(IEnumerable<SelectionRange>? values)
-    {
-        Selection = values?.ToList();
-        await SelectionChanged.InvokeAsync(Selection);
-    }
-
-    /// <summary>
-    /// The active markdown styles at the current selection(s) have changed
-    /// </summary>
-    /// <param name="values"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task MarkdownStyleChangedFromJS(IEnumerable<string>? values)
-    {
-        State.MarkdownStylesAtSelections = values?.ToList() ?? [];
-        await MarkdownStylesAtSelectionsChanged.InvokeAsync(State.MarkdownStylesAtSelections);
-    }
-
-    /// <summary>
-    /// CodeMirror requested linting of the document
-    /// </summary>
-    /// <param name="code"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task<List<CodeMirrorDiagnostic>> LintingRequestedFromJS(string code)
-    {
-        try {
-            LinterCancellationTokenSource.Cancel();
-        }
-        catch (ObjectDisposedException) { }
-        LinterCancellationTokenSource = new();
-        var token = LinterCancellationTokenSource.Token;
-        try {
-            return await LintDocument(code, token);
-        }
-        catch (OperationCanceledException) {
-            return [];
-        }
-    }
-
-    private CancellationTokenSource LinterCancellationTokenSource = new();
-
-    /// <summary>
-    /// A file upload was initiated from the editor
-    /// </summary>
-    /// <param name="fileBytes"></param>
-    /// <param name="fileName"></param>
-    /// <param name="contentType"></param>
-    /// <param name="lastModified"></param>
-    /// <returns></returns>
-    [JSInvokable] public async Task<string?> UploadFileFromJS(byte[] fileBytes, string fileName, string contentType, DateTime lastModified)
-    {
-        using var fileStream = new MemoryStream(fileBytes);
-        var customFormFile = new CustomFormFile(fileStream, fileName, contentType);
-        var customBrowserFile = new CustomBrowserFile(fileStream, fileName, contentType, lastModified);
-        var fileUrl = UploadFile is not null
-            ? await UploadFile(customFormFile)
-            : UploadBrowserFile is not null
-                ? await UploadBrowserFile(customBrowserFile)
-                : null;
-        if (!string.IsNullOrEmpty(fileUrl)) {
-            var imageChar = contentType.StartsWith("image/") ? "!" : string.Empty;
-            var imageLink = $"\n{imageChar}[{fileName}]({fileUrl})\n";
-            //await CmJsInterop!.Commands.Dispatch(CodeMirrorCommandOneParameter.InsertTextAbove, imageLink);
-        }
-        return fileUrl;
-    }
-
-    /*
-    [JSInvokable] public async Task<string> RequestPasteAction(string[] options) => await DoRequestPasteAction(options);
-    */
+    internal CodeMirrorJsInterop? CmJsInterop = null;
+    internal CodeMirrorConfiguration Config = null!;
+    private bool shouldRender = true;
 
     /// <summary>
     /// Life-cycle method invoked when the component is initialized.
